refactor(app): decompose FilterBar into feature folder
- Split the 843-LOC FilterBar.tsx into FilterBar/ with index.tsx
  (orchestrator: scroll/sticky + chip-menu state), FilterSearch.tsx
  (search state, fuse.js results, keyboard navigation, expand/collapse),
  FilterMenu.tsx (category dropdown rendering), FilterChips.tsx (chips
  with roll animation + chip action menu), and FilterSizeToggle.tsx
  (ToolbarActions wrapper)
- Pure decomposition: every sx block, trackEvent call (search,
  search_no_results, grid_resize), aria attribute, and keyboard handler
  copied verbatim; props between parent and children are explicit
  typed interfaces (DropdownItem and RandomAnimation now named types)
- Existing consumers unchanged: index.tsx re-exports FilterBar under
  the same 'src/sections/plots-gallery/FilterBar' path
- Moved FilterBar.test.tsx to FilterBar/index.test.tsx unchanged and
  added per-component tests (keyboard nav, chip menu, dropdown modes,
  size toggle passthrough): 552 -> 578 tests

Part 8 of the frontend modernization roadmap.
