+++ app/src/sections/plots-gallery/FilterBar/index.test.tsx
